bakup/requirements.txt
appdirs==1.4.0
certifi==2017.1.23
cf-deployment-tracker==1.0.2
click==6.7
Flask==0.11.1
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
python-telegram-bot==5.3.0
requests==2.13.0
six==1.10.0
urllib3==1.20
Werkzeug==0.11.15
